requirements.txt

streamlit
pymupdf
fpdf
pillow
